"add admin role , change createdBy and approvedBy data type int to long "

+++ BloggingSite.Models/Entities/ApprovedBlog.cs
@@ -10,7 +10,7 @@ namespace BloggingSite.Models.Entities
 {
     public class ApprovedBlog : PendingBlog
     {       
-        public int ApprovedBy { get; set; }
+        public long ApprovedBy { get; set; }
         public DateTime PublishedDate { get; set; }
         public BlogStatus CurrentStatus { get; set; } = BlogStatus.Create;
